pyproject.toml
[project]
name = "for-cloiud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.47.1",
]
